refactor(Service,Generator): generate 메소드 분리
Service 클래스에서 generate 메소드들을 Generate 클래스로 분리
